+++ bitcoin/src/bip152.rs
@@ -9,11 +9,13 @@ use core::{convert, fmt, mem};
 use std::error;
 
 use hashes::{sha256, siphash24};
-use internals::{impl_array_newtype, ToU64 as _};
+use internals::ToU64 as _;
 use io::{BufRead, Write};
 
 use crate::consensus::encode::{self, Decodable, Encodable, ReadExt, WriteExt};
-use crate::internal_macros::{impl_array_newtype_stringify, impl_consensus_encoding};
+use crate::internal_macros::{
+    impl_array_newtype, impl_array_newtype_stringify, impl_consensus_encoding,
+};
 use crate::prelude::Vec;
 use crate::transaction::TxIdentifier;
 use crate::{block, consensus, Block, BlockHash, Transaction};
